intermediate commit. Now trying to introduce existential SqlValue

+++ src/Delete.hs
@@ -36,10 +36,13 @@ import Action.Tags.Types
 import Action.Authors.Types
 import Action.Category.Types
 import Action.Users.Types
+import Action.Comments.Types
+import ExecuteTypes
+import SqlValue
 
 class (PS.ToRow (Del s)) => DeleteSQL s where
     type Del s :: *
-    deleteQuery :: s -> PS.Query
+    deleteQuery :: s -> Del s -> (PS.Query, [SqlValue])
     dName :: s -> B.ByteString
 
 newtype DTag = DTag ()
@@ -47,7 +50,7 @@ dummyDTag = DTag ()
 
 instance DeleteSQL DTag where
     type Del DTag = DeleteTag
-    deleteQuery _ = "DELETE FROM news.tag WHERE tag_id = ?"
+    deleteQuery _ dt = ("DELETE FROM news.tag WHERE tag_id = ?", [SqlInt $ _dt_tagId dt])
     dName _ = "tag"
 
 
@@ -56,7 +59,7 @@ dummyDCat = DCat ()
 
 instance DeleteSQL DCat where
     type Del DCat = DeleteCategory
-    deleteQuery _ = "DELETE FROM news.category WHERE category_id = ?"
+    deleteQuery _ dc = ("DELETE FROM news.category WHERE category_id = ?", [SqlInt $ _dc_catId dc])
     dName _ = "category"
 
 
@@ -65,7 +68,7 @@ dummyDAuthor = DAuthor ()
 
 instance DeleteSQL DAuthor where
     type Del DAuthor = DeleteAuthor
-    deleteQuery _ = "DELETE FROM news.author WHERE author_id = ?"
+    deleteQuery _ da = ("DELETE FROM news.author WHERE author_id = ?", [SqlInt $ _da_authorId da])
     dName _ = "author"
 
 newtype DUser = DUser ()
@@ -73,8 +76,21 @@ dummyDUser = DUser ()
 
 instance DeleteSQL DUser where
     type Del DUser = DeleteUser
-    deleteQuery _ = "DELETE FROM news.users WHERE user_id = ?"
+    deleteQuery _ du = ("DELETE FROM news.users WHERE user_id = ?", [SqlInt $ _du_userId du])
     dName _ = "user"
 
 
+newtype DComment = DComment ()
+dummyDComment = DComment ()
 
+instance DeleteSQL DComment where
+    type Del DComment = WithUser DeleteComment
+    deleteQuery _ (WithUser u dc) = (
+        "\
+\ DELETE FROM news.comment c \
+\WHERE c.comment_id = ? AND EXISTS \
+\    (SELECT * FROM news.users u WHERE u.user_id = ? AND (u.is_admin = true OR u.user_id = c.user_id))"
+        , [SqlInt $ _dc_commentId dc, SqlInt u])
+
+
+    dName _ = "comment"
